Divided the project into modules
