+++ src/components/films-count.js
@@ -1,0 +1,5 @@
+export const createFilmsCountTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
